(PC-7905): Move AllocineVenueProvider to models file
