+++ src/pcapi/model_creators/generic_creators.py
@@ -13,7 +13,7 @@ from pcapi.core.bookings import api as bookings_api
 from pcapi.core.offerers.models import Offerer
 from pcapi.core.offers.models import Mediation
 from pcapi.core.users.models import TokenType
-from pcapi.core.providers.allocine_venue_provider import AllocineVenueProvider
+from pcapi.core.providers.models import AllocineVenueProvider
 from pcapi.core.providers.models import Provider
 from pcapi.core.providers.models import VenueProvider
 from pcapi.core.users.models import User
