+++ src/pcapi/scripts/add_allocine_internal_ids.py
@@ -4,7 +4,7 @@ import logging
 from sqlalchemy.orm.util import aliased
 
 from pcapi.core.offerers.models import Venue
-from pcapi.core.providers.allocine_venue_provider import AllocineVenueProvider
+from pcapi.core.providers.models import AllocineVenueProvider
 from pcapi.models import db
 from pcapi.models.allocine_pivot import AllocinePivot
