+++ src/pcapi/scripts/venue/modify_allocine_price_rule_for_venue.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 import logging
 
-from pcapi.models import AllocineVenueProvider
+from pcapi.core.providers.models import AllocineVenueProvider
 from pcapi.models import AllocineVenueProviderPriceRule
 from pcapi.models import Venue
 from pcapi.repository import repository
